# packages/api-client/src/generated/openapi.json
{"openapi":"3.1.0","info":{"title":"OpenAPI definition","version":"v0"},"servers":[{"url":"http://localhost:8080","description":"Generated server url"}],"paths":{"/api/v1/spaces":{"get":{"tags":["learning-space-controller"],"operationId":"list","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/LearningSpaceResponse"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["learning-space-controller"],"operationId":"create","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateLearningSpaceRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearningSpaceResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/spaces/{spaceId}/sources":{"get":{"tags":["source-controller"],"operationId":"list_1","parameters":[{"name":"spaceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/SourceResponse"}}}}}},"security":[{"bearerAuth":[]}]},"post":{"tags":["source-controller"],"operationId":"create_1","parameters":[{"name":"spaceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/CreateSourceRequest"}}},"required":true},"responses":{"201":{"description":"Created","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SourceResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/health":{"get":{"tags":["spike-health-controller"],"operationId":"health","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpikeHealthResponse"}}}}}}},"/api/v1/spike/spaces/{spaceId}":{"get":{"tags":["spike-protected-controller"],"operationId":"spaceEndpoint","parameters":[{"name":"spaceId","in":"path","required":true,"schema":{"type":"string"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpikeSpaceAuthorizationResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/spike/protected":{"get":{"tags":["spike-protected-controller"],"operationId":"protectedEndpoint","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpikeStatusResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/spike/method-denied":{"get":{"tags":["spike-protected-controller"],"operationId":"methodDeniedEndpoint","responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SpikeStatusResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/spaces/{spaceId}":{"get":{"tags":["learning-space-controller"],"operationId":"get","parameters":[{"name":"spaceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/LearningSpaceResponse"}}}}},"security":[{"bearerAuth":[]}]}},"/api/v1/spaces/{spaceId}/sources/{sourceId}":{"get":{"tags":["source-controller"],"operationId":"get_1","parameters":[{"name":"spaceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}},{"name":"sourceId","in":"path","required":true,"schema":{"type":"integer","format":"int64"}}],"responses":{"200":{"description":"OK","content":{"application/json":{"schema":{"$ref":"#/components/schemas/SourceResponse"}}}}},"security":[{"bearerAuth":[]}]}}},"components":{"schemas":{"CreateLearningSpaceRequest":{"type":"object","properties":{"name":{"type":"string","maxLength":128,"minLength":0},"description":{"type":"string","maxLength":512,"minLength":0}},"required":["name"]},"LearningSpaceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"name":{"type":"string"},"description":{"type":"string"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"CreateSourceRequest":{"type":"object","properties":{"title":{"type":"string","maxLength":255,"minLength":0},"sourceType":{"type":"string","minLength":1,"pattern":"DESKTOP_UPLOAD|DESKTOP_FOLDER_IMPORT|ADMIN_UPLOAD|ADMIN_MANUAL"}},"required":["sourceType","title"]},"SourceResponse":{"type":"object","properties":{"id":{"type":"integer","format":"int64"},"spaceId":{"type":"integer","format":"int64"},"title":{"type":"string"},"sourceType":{"type":"string"},"status":{"type":"string"},"createdAt":{"type":"string","format":"date-time"},"updatedAt":{"type":"string","format":"date-time"}}},"SpikeHealthResponse":{"type":"object","properties":{"status":{"type":"string"},"service":{"type":"string"},"phase":{"type":"string"}}},"SpikeSpaceAuthorizationResponse":{"type":"object","properties":{"spaceId":{"type":"string"},"status":{"type":"string"}}},"SpikeStatusResponse":{"type":"object","properties":{"status":{"type":"string"}}}},"securitySchemes":{"bearerAuth":{"type":"http","scheme":"bearer","bearerFormat":"JWT"}}}}